+++ client/jetbrains/src/main/kotlin/com/sourcegraph/cody/config/SettingsModel.kt
@@ -5,8 +5,8 @@ import java.awt.Color
 data class SettingsModel(
     var defaultBranchName: String = "",
     var remoteUrlReplacements: String = "",
-    var isCodyEnabled: Boolean = false,
-    var isCodyAutocompleteEnabled: Boolean = false,
+    var isCodyEnabled: Boolean = true,
+    var isCodyAutocompleteEnabled: Boolean = true,
     var isCodyDebugEnabled: Boolean = false,
     var isCodyVerboseDebugEnabled: Boolean = false,
     var isCustomAutocompleteColorEnabled: Boolean = false,